검색페이지 리팩토링 완료
