+++ App.js
@@ -26,9 +26,8 @@ import Login from "./src/Screens/Login";
 import EmailLogin from "./src/Screens/EmailLogin";
 import SignUpID from "./src/Screens/SignUp/ID";
 import SignUpPW from "./src/Screens/SignUp/PW";
-// import Search from "./src/Screens/Search";
-import SearchResult from "./src/Screens/Search2/SearchResult";
-import RecentSearch from "./src/Screens/Search2/RecentSearch";
+import SearchResult from "./src/Screens/Search/SearchResult";
+import RecentSearch from "./src/Screens/Search/RecentSearch";
 import Map from "./src/Screens/Map";
 import MyPage from "./src/Screens/MyPage";
 import KakaoLogin from "./src/Screens/Social/Login/Kakao/index";
