+++ app.json
@@ -28,7 +28,11 @@
       "adaptiveIcon": {
         "foregroundImage": "./assets/adaptive-icon.png",
         "backgroundColor": "#FFFFFF"
+      },
+      "config": {
+        "googleMaps": {"apiKey": "AIzaSyAS9tWbUhSQZnx-2_as1gBAKK5GIPhkI6g"}
       }
+
     },
     "web": {
       "favicon": "./assets/favicon.png"
